Add public contributor profiles
Now anyone can visit postgresql.org/community/contributors/[username] to
see a user's self-described contributions. This commit adds the profile
view and also links recognized contributors' names to their profiles.

The urls are based on username, which may not be the right design
because not all contributors have user profiles, but it is a starting
point.