Make public contributor profile opt-in
I'm not really sure if this is the right way to do this at all, so I
made it a separate patch. Basically, we probably don't want to just
suddenly have everybody with a user account at postgresql.org have a
public profile page. But, I don't know how it should work exactly --
both from a feature specification perspective and an implementation
perspective.

For one, we can't have all existing contributors not be displayed on the
recognized contributor page just because they didn't opt-in (which is
what I did here). But do we want all of them to get profiles without
first opting in?